# crates/birdnet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BirdNet-Behavior HTTP API",
    "version": "0.16.0",
    "summary": "JSON API for a BirdNet-Behavior acoustic monitoring station: open read-only endpoints, plus a small token-authenticated write surface.",
...
    "license": {
      "name": "CC BY-NC-SA 4.0",
      "url": "https://github.com/tomtom215/BirdNet-Behavior/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "Versioned API root, relative to the station's base URL (default http://localhost:8502)."
    }
  ],
  "security": [],
...
  "paths": {
...
    "/species/image/{scientific_name}": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Species image metadata",
        "description": "Metadata for a cached Wikipedia/Wikimedia image of the species (URL, attribution, cache status). Fetches and caches on a miss.",
        "operationId": "speciesImageInfo",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScientificName"
          }
        ],
        "responses": {
          "200": {
            "description": "Image metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesImage"
                }
              }
            }
          }
        }
      }
    },
    "/species/image/{scientific_name}/file": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Species image bytes",
        "description": "The cached image itself. Checks a configured custom image directory first, then the Wikipedia cache.",
        "operationId": "speciesImageFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScientificName"
          }
        ],
        "responses": {
          "200": {
            "description": "Image bytes.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No cached image available."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The station's `BNB_API_TOKEN`, resolved from its configuration file and then its environment. Required by every write endpoint, and by the settings read — a station's configuration is not public even with its credentials taken out. The read-only detection, species and analytics endpoints accept no credential and ignore one if it is sent."
      }
    },
    "parameters": {
...
      "ScientificName": {
        "name": "scientific_name",
        "in": "path",
        "required": true,
        "description": "Scientific name of the species.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "schemas": {
...
      "SpeciesImage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "scientific_name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "cached": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "wiki_url": {
            "type": "string"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
...
    }
  }
}